+++ components/account/ProfileSettingsCard.tsx
@@ -1,13 +1,13 @@
-import { Card, CardContent, Typography, } from '@mui/material';
+import { Card, CardContent, Typography } from '@mui/material';
 import { useTranslations } from 'next-intl';
-import ProfileDetailsForm from '../forms/ProfileDetailsForm';
+import ProfileSettingsForm from '../forms/ProfileSettingsForm';
 
 const formCardStyle = {
   width: { xs: '100%', sm: '70%', md: '45%' },
   alignSelf: 'flex-start',
 } as const;
 
-const ProfileSettings = () => {
+const ProfileSettingsCard = () => {
   const t = useTranslations('Account.accountSettings');
 
   return (
@@ -16,13 +16,11 @@ const ProfileSettings = () => {
         <Typography variant="h2" component="h2">
           {t('profile.title')}
         </Typography>
-        <Typography fontSize="1rem !important">
-          {t('profile.description')}
-        </Typography>
-        <ProfileDetailsForm />
+        <Typography fontSize="1rem !important">{t('profile.description')}</Typography>
+        <ProfileSettingsForm />
       </CardContent>
     </Card>
-  )
-}
+  );
+};
 
-export default ProfileSettings
+export default ProfileSettingsCard;
